Add a stubbed out registration process.
Works likes this:

If a user goes to /pages/register.html there are presented with a form
that lives in templates/register.html that askes them for their name,
user, company.

That information is emailed to the address named in tiddlywebconfig
wikidata.register_address (the default is currently me, this should change!).

Once the email is sent the user is told they will be contacted later.

That's it for now.

The next step is to provide a user creation form for the admins.