Add /_admin/updateuser' with GET=update_user_form, POST=update_user
To use

Go to /_admin/updateuser?username=<someusername>

This will load a form containing the current user info for
<someusername>. Upon submission the information is updated.

A side effect allows this form to be used to create users as
well if that's desired.

An interface is needed to properly call the form.

The form could be tuned up quite a bit to reflect a better UI
for setting expiration times and tier settings.