If the extractor has detetected that a user has expired handle it properly.
When someone expires, the extract sets expired_user but the
user retains their cookie so they will get expired_user over and
over. This change sets it so that when they are sent to the login
form, they will be logged out as part of that request, iff they
are expired.