First attempt at getting searching of branches handled.
This creates a branches boolean based on the incoming query
string, and then checks that to see whether some entity_type
filters are to be added to the sql query.

This change introduces a great deal of duplication between
madsql and wdsql which will be resolved in a later commit
once this stuff works.