Automate handling of data imports a bit more
This is still not the full suite, but does speed
things up a bit:

If the name of the imported file contains 'wiki_full', delete
all the rows from the table, and do a full import. Otherwise
just do a a delta import.

One the import is done, automatically update the fruitmachine,
which now looks like this:

<span id="fruitmachine" class="lightborder darkgreen">
    {% set counter = '322618' %}
    {% for i in counter %}
    <img src="/static/test_images/count{{ i }}.png" alt="{{ i }}" />
    {% endfor %}
</span>

The import script changes the value of counter.