Add rudimentary support for a v2 search that sorts by relevance.
There are other features that still need to be added, this just gets
the relevancy handling in place.

This requires the addition of another fulltext index on just the legal_name
field so that it can be used to weight the relevancy returns. The assumption
is that if there is a better match in legal_name than in previous_name then
that should be reflected in the ordering of the results.

It's an open question still whether the revelancy ranking should be
by boolean mode or natural language (the default). The latter does
some word frequency counting to ignore terms like 'the', which might
not be a good idea for us. However, the boolean mode only returns
1 on a match, not a float, so the relevancy doesn't really sort...

This commit also introduces different quote handling for v2. If
there are no quotes present than the terms are considered separate,
if there are quotes, then that quoted thing is considered a phrase.

In v1 a ' ' causes quoting.