Add support for a type query parameter that controls how searches are performed
type can be all, exact or partial (e.g. type=all)

given the query string of london bank

if all:
    +london +bank # return anything with both london and query

if exact:
    "london bank" # return matches of the phrase

if partial:

    london* bank* # contains either london or bank or also
         londinium or banking

To make this go the search interface in the templates+js needs to be
updated so that the query string is set with a v=2 and a query=all,
or query=exact or query=partial.