Adjust searching on an avid to ignore branch restrictions:
"A search against an AVID should not refer to the value of the Include
Branches checkbox. The associated record should always be returned
regardless of whether it is a branch or not. So, if an AVID is entered
into the text string, assume that including branches is mandatory."

Made an adjustment to miniwiki.dump so that an entry is a branch.