Add support for generating a search query which list all entities with a p(lei).
If a query is formed like so:

   /search?v=2;q=;lei=all

Then the results will be those entities which have an lei, modulo whatever
the slice index is set to be.

The code for this is a bit rough, but does the job. It's not as tidy as it
should be because a) this was a last minute request, b) there are
several layers of cruft in the containing method, which is a generator,
which makes it harder to extract the necessary bits.